+++ zarrs_metadata_ext/CHANGELOG.md
@@ -10,4 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Split from the `zarrs_metadata` module of `zarrs_metadata` 0.4.0
 
+### Changed
+- rename `ArrayMetadataV2ToV3ConversionError` to `ArrayMetadataV2ToV3Error`
+- rename `InvalidPermutationError` to `TransposeOrderError`
+
 [unreleased]: https://github.com/zarrs/zarrs/compare/zarrs_metadata-v0.1.0...HEAD